requirements.txt
beautifulsoup4==4.12.2
feedparser==6.0.10
newspaper3k==0.2.8
requests==2.31.0
openai==1.0.0
langchain==0.0.268
sqlalchemy==2.0.19
reportlab==4.0.5
python-dotenv==1.0.0
pydantic==2.2.1
uvicorn==0.23.2
fastapi==0.95.2
scikit-learn==1.3.2
tqdm==4.66.1
PyYAML==6.0
